Test saving POS models and regroup mapping tests
