+++ Modules/Tests/YosemiteTests/Storage/PersistedProductVariationTests.swift
@@ -1,0 +1,335 @@
+import Foundation
+import Testing
+import GRDB
+@testable import Yosemite
+@testable import Storage
+
+struct PersistedProductVariationTests {
+
+    @Test("PersistedProductVariation init(from:) maps all POSProductVariation fields")
+    func test_variation_init_from_posProductVariation_maps_all_fields() throws {
+        // Given
+        let siteID: Int64 = 5
+        let productID: Int64 = 50
+        let variationID: Int64 = 500
+        let attrs = [
+            ProductVariationAttribute(id: 0, name: "Color", option: "Green"),
+            ProductVariationAttribute(id: 0, name: "Size", option: "XL")
+        ]
+        let image = ProductImage(imageID: 501,
+                                 dateCreated: Date(timeIntervalSince1970: 1000),
+                                 dateModified: nil,
+                                 src: "https://example.com/v.png",
+                                 name: "v",
+                                 alt: nil)
+        let pos = POSProductVariation(
+            siteID: siteID,
+            productID: productID,
+            productVariationID: variationID,
+            attributes: attrs,
+            image: image,
+            fullDescription: "VFull",
+            sku: "VSKU",
+            globalUniqueID: "VGID",
+            price: "19.95",
+            downloadable: false,
+            manageStock: true,
+            stockQuantity: 2,
+            stockStatusKey: "instock"
+        )
+
+        // When
+        let persisted = PersistedProductVariation(from: pos)
+
+        // Then
+        #expect(persisted.id == variationID)
+        #expect(persisted.siteID == siteID)
+        #expect(persisted.productID == productID)
+        #expect(persisted.sku == pos.sku)
+        #expect(persisted.globalUniqueID == pos.globalUniqueID)
+        #expect(persisted.price == pos.price)
+        #expect(persisted.downloadable == pos.downloadable)
+        #expect(persisted.fullDescription == pos.fullDescription)
+        #expect(persisted.manageStock == pos.manageStock)
+        #expect(persisted.stockQuantity == pos.stockQuantity)
+        #expect(persisted.stockStatusKey == pos.stockStatusKey)
+    }
+
+    @Test("PersistedProductVariation toPOSProductVariation maps back with attributes and optional image")
+    func test_variation_toPOSProductVariation_maps_back_including_attributes_and_image() throws {
+        // Given
+        let siteID: Int64 = 6
+        let productID: Int64 = 60
+        let variationID: Int64 = 600
+        let persisted = PersistedProductVariation(
+            id: variationID,
+            siteID: siteID,
+            productID: productID,
+            sku: "SKU",
+            globalUniqueID: "GID",
+            price: "11.00",
+            downloadable: true,
+            fullDescription: "Full",
+            manageStock: false,
+            stockQuantity: nil,
+            stockStatusKey: "outofstock"
+        )
+
+        let varAttrs = [
+            PersistedProductVariationAttribute(productVariationID: variationID, name: "Material", option: "Wool"),
+            PersistedProductVariationAttribute(productVariationID: variationID, name: "Fit", option: "Slim")
+        ]
+        let varImage = PersistedProductVariationImage(
+            id: 601,
+            productVariationID: variationID,
+            dateCreated: Date(timeIntervalSince1970: 2000),
+            dateModified: Date(timeIntervalSince1970: 3000),
+            src: "https://example.com/vi.png",
+            name: "vi",
+            alt: "vai")
+
+        // When
+        let pos = persisted.toPOSProductVariation(
+            attributes: varAttrs.map { $0.toProductVariationAttribute() },
+            image: varImage.toProductImage()
+        )
+
+        // Then
+        #expect(pos.siteID == siteID)
+        #expect(pos.productID == productID)
+        #expect(pos.productVariationID == variationID)
+        #expect(pos.sku == persisted.sku)
+        #expect(pos.globalUniqueID == persisted.globalUniqueID)
+        #expect(pos.price == persisted.price)
+        #expect(pos.downloadable == persisted.downloadable)
+        #expect(pos.fullDescription == persisted.fullDescription)
+        #expect(pos.manageStock == persisted.manageStock)
+        #expect(pos.stockQuantity == persisted.stockQuantity)
+        #expect(pos.stockStatusKey == persisted.stockStatusKey)
+        #expect(pos.attributes.count == 2)
+        #expect(pos.image?.imageID == varImage.id)
+    }
+
+    @Test("ProductVariation with associations fetches attributes and image automatically")
+    func test_product_variation_with_associations_fetches_related_records() throws {
+        let grdbManager = try GRDBManager()
+        let db = grdbManager.databaseConnection
+
+        try db.write { db in
+            // Insert required site and product first
+            let site = PersistedSite(id: 2)
+            try site.insert(db)
+
+            let parentProduct = PersistedProduct(
+                id: 50,
+                siteID: 2,
+                name: "Parent Product",
+                productTypeKey: "variable",
+                fullDescription: nil,
+                shortDescription: nil,
+                sku: nil,
+                globalUniqueID: nil,
+                price: "0.00",
+                downloadable: false,
+                parentID: 0,
+                manageStock: false,
+                stockQuantity: nil,
+                stockStatusKey: "instock"
+            )
+            try parentProduct.insert(db)
+
+            // Insert test data
+            let variation = PersistedProductVariation(
+                id: 500,
+                siteID: 2,
+                productID: 50,
+                sku: "VAR-SKU",
+                globalUniqueID: "GID-500",
+                price: "15.50",
+                downloadable: true,
+                fullDescription: "Variation description",
+                manageStock: false,
+                stockQuantity: nil,
+                stockStatusKey: "outofstock"
+            )
+            try variation.insert(db)
+
+            let variationImage = PersistedProductVariationImage(
+                id: 600,
+                productVariationID: 500,
+                dateCreated: Date(timeIntervalSince1970: 3000),
+                dateModified: Date(timeIntervalSince1970: 3500),
+                src: "https://example.com/var-img.png",
+                name: "Variation Image",
+                alt: "Variation alt text"
+            )
+            try variationImage.insert(db)
+
+            var attr1 = PersistedProductVariationAttribute(
+                productVariationID: 500,
+                name: "Color",
+                option: "Purple"
+            )
+            var attr2 = PersistedProductVariationAttribute(
+                productVariationID: 500,
+                name: "Material",
+                option: "Cotton"
+            )
+            try attr1.insert(db)
+            try attr2.insert(db)
+        }
+
+        // Test automatic fetching via associations
+        let fetchedVariation = try db.read { db in
+            try PersistedProductVariation.filter(PersistedProductVariation.Columns.id == 500).fetchOne(db)
+        }
+
+        let variation = try #require(fetchedVariation)
+        let posVariation = try variation.toPOSProductVariation(db: db)
+
+        // Verify variation fields
+        #expect(posVariation.productVariationID == 500)
+        #expect(posVariation.siteID == 2)
+        #expect(posVariation.productID == 50)
+        #expect(posVariation.sku == "VAR-SKU")
+        #expect(posVariation.fullDescription == "Variation description")
+        #expect(posVariation.price == "15.50")
+        #expect(posVariation.downloadable == true)
+        #expect(posVariation.manageStock == false)
+        #expect(posVariation.stockQuantity == nil)
+        #expect(posVariation.stockStatusKey == "outofstock")
+
+        // Verify image was fetched
+        #expect(posVariation.image?.imageID == 600)
+        #expect(posVariation.image?.src == "https://example.com/var-img.png")
+        #expect(posVariation.image?.name == "Variation Image")
+        #expect(posVariation.image?.alt == "Variation alt text")
+
+        // Verify attributes were fetched
+        #expect(posVariation.attributes.count == 2)
+        let colorAttr = posVariation.attributes.first { $0.name == "Color" }
+        #expect(colorAttr?.option == "Purple")
+
+        let materialAttr = posVariation.attributes.first { $0.name == "Material" }
+        #expect(materialAttr?.option == "Cotton")
+    }
+
+    @Test("ProductVariation without image returns nil image")
+    func test_product_variation_without_image_returns_nil() throws {
+        let grdbManager = try GRDBManager()
+        let db = grdbManager.databaseConnection
+
+        try db.write { db in
+            // Insert required site and product first
+            let site = PersistedSite(id: 4)
+            try site.insert(db)
+
+            let parentProduct = PersistedProduct(
+                id: 70,
+                siteID: 4,
+                name: "Parent Product",
+                productTypeKey: "variable",
+                fullDescription: nil,
+                shortDescription: nil,
+                sku: nil,
+                globalUniqueID: nil,
+                price: "0.00",
+                downloadable: false,
+                parentID: 0,
+                manageStock: false,
+                stockQuantity: nil,
+                stockStatusKey: "instock"
+            )
+            try parentProduct.insert(db)
+
+            // Insert variation with attributes but no image
+            let variation = PersistedProductVariation(
+                id: 700,
+                siteID: 4,
+                productID: 70,
+                sku: "VAR-NO-IMG",
+                globalUniqueID: "GID-700",
+                price: "12.00",
+                downloadable: false,
+                fullDescription: "No image variation",
+                manageStock: true,
+                stockQuantity: 3,
+                stockStatusKey: "instock"
+            )
+            try variation.insert(db)
+
+            // Add attributes only, no image
+            var attr = PersistedProductVariationAttribute(
+                productVariationID: 700,
+                name: "Style",
+                option: "Modern"
+            )
+            try attr.insert(db)
+        }
+
+        // Test that variation with no image returns nil for image
+        let fetchedVariation = try db.read { db in
+            try PersistedProductVariation.filter(PersistedProductVariation.Columns.id == 700).fetchOne(db)
+        }
+
+        let variation = try #require(fetchedVariation)
+        let posVariation = try variation.toPOSProductVariation(db: db)
+
+        #expect(posVariation.productVariationID == 700)
+        #expect(posVariation.image == nil)
+        #expect(posVariation.attributes.count == 1)
+        #expect(posVariation.attributes.first?.name == "Style")
+        #expect(posVariation.attributes.first?.option == "Modern")
+    }
+
+    @Test("PersistedProductVariationAttribute init(from:) and toProductVariationAttribute round-trip")
+    func test_variation_attribute_round_trip() throws {
+        // Given
+        let variationID: Int64 = 700
+        let attr = ProductVariationAttribute(id: 0, name: "Style", option: "Modern")
+
+        // When
+        let persisted = PersistedProductVariationAttribute(from: attr, productVariationID: variationID)
+        let back = persisted.toProductVariationAttribute()
+
+        // Then
+        #expect(persisted.productVariationID == variationID)
+        #expect(persisted.name == attr.name)
+        #expect(persisted.option == attr.option)
+
+        #expect(back.name == attr.name)
+        #expect(back.option == attr.option)
+    }
+
+    @Test("PersistedProductVariationImage init(from:) and toProductImage round-trip")
+    func test_variation_image_round_trip() throws {
+        // Given
+        let variationID: Int64 = 800
+        let image = ProductImage(imageID: 801,
+                                 dateCreated: Date(timeIntervalSince1970: 4000),
+                                 dateModified: nil,
+                                 src: "https://example.com/img.png",
+                                 name: nil,
+                                 alt: nil)
+
+        // When
+        let persisted = PersistedProductVariationImage(from: image, productVariationID: variationID)
+        let back = persisted.toProductImage()
+
+        // Then
+        #expect(persisted.id == image.imageID)
+        #expect(persisted.productVariationID == variationID)
+        #expect(persisted.dateCreated == image.dateCreated)
+        #expect(persisted.dateModified == image.dateModified)
+        #expect(persisted.src == image.src)
+        #expect(persisted.name == image.name)
+        #expect(persisted.alt == image.alt)
+
+        #expect(back.imageID == image.imageID)
+        #expect(back.dateCreated == image.dateCreated)
+        #expect(back.dateModified == image.dateModified)
+        #expect(back.src == image.src)
+        #expect(back.name == image.name)
+        #expect(back.alt == image.alt)
+    }
+}
